Fixed type miss match and name changes in entities to fix build and compile errors. Regenerated entity and session beans after database changes.

+++ mtams/mtams-war/src/java/ManagedBeans/ApprovalBean.java
@@ -215,7 +215,7 @@ public class ApprovalBean implements Serializable {
             appr.setNotes(approvalComment);
             approvalHandler.persistApproval(appr);
             
-            short s = 1;
+            int s = 1;
             selectedAccQte.setSelected(s);
             selectedCarQte.setSelected(s);
             selectedFlgQte.setSelected(s);
